backend/requirements.txt
absl-py==1.0.0
aioice==0.9.0
aioredis==1.3.1
aiortc==1.9.0
amqp==5.0.9
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
autopep8==1.6.0
av==12.3.0
bcrypt==3.2.0
billiard==3.6.4.0
cachetools==4.2.4
celery==5.2.2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==23.10.4
# cryptography==43.0.3
cycler==0.11.0
daphne==3.0.2
Django==4.2.16
django-celery-results==2.2.0
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
dnspython==2.1.0
ffmpeg-python==0.2.0
flatbuffers==2.0
fonttools==4.28.5
future==0.18.2
gast==0.4.0
gevent==24.11.1
greenlet==3.1.1
grpcio==1.43.0
gunicorn==23.0.0
h11==0.12.0
# h5py==3.6.0
hiredis==3.0.0
hyperlink==21.0.0
idna==3.3
ifaddr==0.2.0
importlib-metadata==4.8.2
incremental==24.7.2
# keras==2.8.0
# Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kombu==5.2.2
libclang==12.0.0
Markdown==3.3.6